+++ taichi_three/__init__.py
@@ -26,6 +26,7 @@ from .buffer import *
 from .loader import *
 from .light import *
 from .raycast import *
+from .skybox import *
 from .objedit import *
 
 print(f'[Tai3D] Camera control hints: LMB to orbit, MMB to move, RMB to scale')